% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-age.R
\name{ageFoI}
\alias{ageFoI}
\title{Functions to modify the FoI by age}
\usage{
ageFoI(a, par)
}
\arguments{
\item{a}{host age}

\item{par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
This method dispatches on \code{par}.
}
