% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-age.R
\name{ageFoI.type2}
\alias{ageFoI.type2}
\title{A function that does not modify the FoI by age}
\usage{
\method{ageFoI}{type2}(a, par)
}
\arguments{
\item{a}{host age}

\item{par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
A function that does not modify the FoI by age
}
