% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{cdfConvolve2}
\alias{cdfConvolve2}
\title{The density function for the sum of two infections}
\usage{
cdfConvolve2(meshX, CDF1, CDF2)
}
\arguments{
\item{meshX}{a mesh of host cohort ages}

\item{CDF1}{the CDF for a complex distribution}

\item{CDF2}{the CDF for a complex distribution}
}
\value{
a \link{numeric} vector of length(meshX)
}
\description{
The density function for the sum of two infections
}
