% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{counts_MLE}
\alias{counts_MLE}
\title{Compute the MLE for parasite counts data}
\usage{
counts_MLE(alpha, counts, inits, sz, F_b, F_Sa, q = 6, bvm = 13)
}
\arguments{
\item{alpha}{the ages of infection}

\item{counts}{the counts}

\item{inits}{the initial guesses for the parameters}

\item{sz}{the size parameter}

\item{F_b}{a function to constrain the upper bound for fitting for \link{Fmu}}

\item{F_Sa}{a function to constrain the slope for fitting for \link{Fmu}}

\item{q}{the sampling volume (in \eqn{\log_{10} microliters})}

\item{bvm}{the blood volume (in \eqn{\log_{10} microliters})}
}
\value{
the fitted parameter values and the log likelihood
}
\description{
Compute the MLE for parasite counts data
}
