% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2density.R
\name{d_alpha2density}
\alias{d_alpha2density}
\title{The density function for parasite densities in a simple malaria infection of age alpha}
\usage{
d_alpha2density(
  xi,
  alpha,
  W = 0,
  a = 0,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_Omega = par_Omega_beta()
)
}
\arguments{
\item{xi}{a vector of quantiles}

\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{a}{host cohort age}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{Fmu}}

\item{par_Omega}{parameters to compute \link{d_Omega}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
The density function for parasite densities in a simple malaria infection of age alpha
}
