% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{d_clone_detect}
\alias{d_clone_detect}
\title{Detection of infection given parasitemia}
\usage{
d_clone_detect(
  a,
  FoI_a,
  bday = 0,
  hhat = 1,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  par_Omega = par_Omega_beta(),
  pWda = par_Wda_none(),
  par_sample = par_nb()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{Fmu}}

\item{par_Omega}{parameters to compute \link{d_Omega}}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{par_sample}{parameters that define a detection function}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
Detection of infection given parasitemia
}
