% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-nb.R
\name{d_counts.nb}
\alias{d_counts.nb}
\title{Negative binomial PDF for raw parasite counts}
\usage{
\method{d_counts}{nb}(x, xi, bvm = 13, pars = par_nb())
}
\arguments{
\item{x}{raw parasite counts}

\item{xi}{mean \eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{pars}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Negative binomial PDF for raw parasite counts
}
