% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-interface.R
\name{d_counts}
\alias{d_counts}
\title{The PDF for counting \eqn{x} parasites}
\usage{
d_counts(x, xi, bvm = 13, pars = par_nb())
}
\arguments{
\item{x}{raw parasite counts}

\item{xi}{mean \eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{pars}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
The PDF for counting \eqn{x} parasites
}
