% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pois.R
\name{d_detect.pois}
\alias{d_detect.pois}
\title{Detection of infection given parasitemia}
\usage{
\method{d_detect}{pois}(xi, bvm = 13, pars = par_pois())
}
\arguments{
\item{xi}{\eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{pars}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
