% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-interface.R
\name{d_detect}
\alias{d_detect}
\title{The probability of detecting parasites}
\usage{
d_detect(xi, bvm = 13, pars = par_nb())
}
\arguments{
\item{xi}{\eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{pars}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
The probability of detecting parasites
}
