% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_parasites.R, R/detection.R
\name{d_detect_mesh}
\alias{d_detect_mesh}
\title{Detection of infection given parasitemia}
\usage{
d_detect_mesh(xi_mesh, xi_density, bvm = 13, par_sample = par_nb())

d_detect_mesh(xi_mesh, xi_density, bvm = 13, par_sample = par_nb())
}
\arguments{
\item{xi_mesh}{a mesh over \eqn{\log_{10}} parasite densities}

\item{xi_density}{the density distribution for parasites over xi_mesh}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{par_sample}{parameters that define a detection function}
}
\value{
the fraction of infected hosts that would test positive

the fraction of infected hosts that would test positive
}
\description{
Detection of infection given parasitemia

Detection of infection given parasitemia
}
