% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-interface.R
\name{d_nz_counts_log}
\alias{d_nz_counts_log}
\title{The PDF for non-zero counts, \eqn{\log_{10}} transformed}
\usage{
d_nz_counts_log(hatxi, xi, bvm = 13, pars = par_nb())
}
\arguments{
\item{hatxi}{\eqn{\log_{10}} parasite counts}

\item{xi}{\eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{pars}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
The PDF for non-zero counts, \eqn{\log_{10}} transformed
}
