% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-beta.R
\name{d_Omega.beta}
\alias{d_Omega.beta}
\title{Modified beta distribution, density function}
\usage{
\method{d_Omega}{beta}(xi, mu, bvm = 13, par_Omega = par_Omega_beta())
}
\arguments{
\item{xi}{vector of quantiles for \eqn{\log_{10}} parasite densities}

\item{mu}{the expected value for \eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{par_Omega}{parameters to compute parasite densities as a function of mu}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The beta distribution, parameterized by the mean and variance, modified to return
a number between 0 and bvm (the number of red blood cells)
}
