% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_parasites.R
\name{d_parasite_density}
\alias{d_parasite_density}
\title{Compute the distribution function for \eqn{B_\bday(a)}}
\usage{
d_parasite_density(
  meshX,
  a,
  FoI_a,
  bday = 0,
  hhat = 1,
  r = 1/200,
  RBC_par = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  Omega_par = par_Omega_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{meshX}{a mesh over parasite densities}

\item{a}{host cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{RBC_par}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{Fmu}}

\item{Omega_par}{parameters to compute \link{d_Omega}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} vector of length meshX
}
\description{
Call \link{parasite_density} and return the PDF:
\deqn{f_B(\xi; a, \bday |h) = \log_{10} \left( \sum_{M_\bday(a|h)>0} 10^{P_\bday(a |h)}\right)}
}
