% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{dAoI}
\alias{dAoI}
\title{Compute the density function for AoI}
\usage{
dAoI(alpha, a, FoI_a, bday = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{alpha}{the age of an infection, \eqn{\alpha}}

\item{a}{cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
The density of the AoI is given by
\deqn{f_A(\alpha | a, \bday, h) = \frac{\int_0^a z(\alpha, a)}{m_\bday(a)}}
}
