% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{dAoYda}
\alias{dAoYda}
\title{Compute the derivatives of the approximate moments of the AoY dynamically}
\usage{
dAoYda(a, vars, pars, FoI_a)
}
\arguments{
\item{a}{the host age}

\item{vars}{the state variables}

\item{pars}{the parameters}

\item{FoI_a}{a cohort trace function}
}
\value{
the derivatives of the MoI and the AoI as a \link{list}
}
\description{
The
}
