% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{dDensityPaConvolve2}
\alias{dDensityPaConvolve2}
\title{The density function for the sum of two infections}
\usage{
dDensityPaConvolve2(
  x,
  a,
  FoI_a,
  hhat = NULL,
  bday = 0,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_Omega = par_Omega_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{x}{host cohort age}

\item{a}{host cohort age}

\item{FoI_a}{a cohort trace function}

\item{hhat}{a local scaling parameter for the FoI}

\item{bday}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{Fmu}}

\item{par_Omega}{parameters to compute \link{d_Omega}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} value
}
\description{
The density function for the sum of two infections
}
