% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{dmda}
\alias{dmda}
\title{Compute the derivatives for MoI using a hybrid model}
\usage{
dmda(a, M, p, FoI_a)
}
\arguments{
\item{a}{the host age}

\item{M}{the state variables}

\item{p}{the parameters}

\item{FoI_a}{a cohort trace function}
}
\value{
the derivatives as a \link{list}
}
\description{
Compute the derivatives for MoI using a hybrid model
}
