% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{dpda}
\alias{dpda}
\title{Compute the derivatives for MoI and true PR}
\usage{
dpda(a, M, par, FoI_a)
}
\arguments{
\item{a}{the host age}

\item{M}{state variables}

\item{par}{the model parameters}

\item{FoI_a}{a cohort trace function}
}
\value{
the derivatives, as a \link{list}
}
\description{
Compute the derivatives for MoI and true PR
}
