% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquIPz.R
\name{dSquIPz}
\alias{dSquIPz}
\title{Derivatives for \link{SquIP}}
\usage{
dSquIPz(a, y, pars, FoIpar, Z)
}
\arguments{
\item{a}{the host age}

\item{y}{a vector of state variables}

\item{pars}{the parameters}

\item{FoIpar}{\eqn{h_\tau(a)}, a \link{list} formatted to compute \link{FoI}}

\item{Z}{the MoE matrix}
}
\value{
the derivatives as a \link{list}
}
\description{
This computes the derivatives for the queuing model \link{SquIP}.
It also computes the hybrid variables \eqn{I, m_1} and \eqn{m_2}
}
\seealso{
\link{SquIP} & \link{solve_SquIP}

This function computes the derivatives in a form that can be used by \link[deSolve:ode]{deSolve::ode}.

\link{solveMMinfty}
}
