% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoyN.R
\name{dYoN}
\alias{dYoN}
\title{The youngest of N infections, density function}
\usage{
dYoN(N, a, FoI_a, hhat = NULL, bday = 0, r = 1/200)
}
\arguments{
\item{N}{the number of infections}

\item{a}{the age of a cohort}

\item{FoI_a}{a cohort trace function}

\item{hhat}{a local scaling parameter for the FoIs}

\item{bday}{the cohort birthday}

\item{r}{the clearance rate of a simple infection}
}
\value{
a \link{numeric} vector of length a + 1
}
\description{
The youngest of N infections, density function
}
