% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmu-immune.R
\name{Fmu.W}
\alias{Fmu.W}
\title{Compute expected log10 parasite densities, `mu`, as a function of the age of infection `alpha`}
\usage{
\method{Fmu}{W}(alpha, W, par_Fmu)
}
\arguments{
\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{par_Fmu}{a [list] that defines a model}
}
\value{
mean log10 parasite densities
}
\description{
Compute mean log10 parasites for a model with no immunity
}
