% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi.R
\name{FoI}
\alias{FoI}
\title{Force of Infection, \eqn{h_\tau(a)}}
\usage{
FoI(a, FoIpar, tau = 0)
}
\arguments{
\item{a}{host age}

\item{FoIpar}{a compound \link{list} to compute \eqn{h_\tau(a)}}

\item{tau}{the host cohort's birthday}
}
\value{
\link{numeric}
}
\description{
Given a birth day, \eqn{\tau,} age is related
to time by the formula \eqn{t = a+\tau}.
The FoI is computed as \deqn{\bar h \; \omega(a) S(a+\tau) T(a+\tau)}
}
