% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{llik_count}
\alias{llik_count}
\title{Compute the likelihood for a single parasite count datum}
\usage{
llik_count(count, mu, sz, par_O, q = 6, bvm = 13)
}
\arguments{
\item{count}{the parasite count}

\item{mu}{the expected value for the count, given age}

\item{sz}{the size parameter for a negative binomial distribution}

\item{par_O}{parameters \link{d_Omega}}

\item{q}{the sampling volume (in \eqn{\log_{10} microliters})}

\item{bvm}{the blood volume (in \eqn{\log_{10} microliters})}
}
\value{
the negative log likelihood for the observation
}
\description{
Compute the likelihood for a single parasite count datum
}
