% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{llik_counts_sz}
\alias{llik_counts_sz}
\title{Compute the log likelihood for a set of parasites counts data}
\usage{
llik_counts_sz(
  prms,
  sz,
  alpha,
  counts,
  par_Fmu,
  par_O,
  F_b,
  F_Sa,
  q = 6,
  bvm = 13
)
}
\arguments{
\item{prms}{A set of parameters to be fitted}

\item{sz}{the size parameter}

\item{alpha}{the AoI for the observations}

\item{counts}{the counts}

\item{par_Fmu}{an object to dispatch \link{Fmu}}

\item{par_O}{an object to dispath \link{d_Omega}}

\item{F_b}{a function to constrain the upper bound for fitting for \link{Fmu}}

\item{F_Sa}{a function to constrain the slope for fitting for \link{Fmu}}

\item{q}{the sampling volume (in \eqn{\log_{10} microliters})}

\item{bvm}{the blood volume (in \eqn{\log_{10} microliters})}
}
\value{
the negative log likelihood for a set of observations
}
\description{
Compute the log likelihood for a set of parasites counts data
}
