% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBC.R
\name{log10RBC}
\alias{log10RBC}
\title{Compute log10 of the red blood cell population}
\usage{
log10RBC(a, par)
}
\arguments{
\item{a}{the age of a host cohort}

\item{par}{a [list] of parameters}
}
\value{
a [numeric] vector of length(a)
}
\description{
Compute log10 of the red blood cell population
}
