% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBC.R
\name{log10RBC.static}
\alias{log10RBC.static}
\title{Compute log10 of the red blood cell population}
\usage{
\method{log10RBC}{static}(a, par)
}
\arguments{
\item{a}{the age of a host cohort}

\item{par}{a \link{list} of parameters}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
A static model
}
