% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_dynamics.R
\name{meanMoI}
\alias{meanMoI}
\title{The mean MoI in a host cohort of age \eqn{a}}
\usage{
meanMoI(a, FoI_a, bday = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{a}{cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
The mean multiplicity of infection (MoI) is \deqn{m_\bday(a|h) = \int_0^a z_\bday(\alpha, a|h) d \alpha}
}
