% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_clone.R
\name{moments_clone_density}
\alias{moments_clone_density}
\title{Compute the moments of P_density}
\usage{
moments_clone_density(
  a,
  FoI_a,
  bday = 0,
  n = 1,
  dt = 0.1,
  hhat = 1,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_Omega = par_Omega_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{n}{the moment to compute}

\item{dt}{the mesh size over xi}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{Fmu}}

\item{par_Omega}{parameters to compute \link{d_Omega}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
Compute \deqn{P_\bday(a| h) \sim f_P(\xi; a, \bday |h ) = \int_0^a \Omega(\xi|F_\mu(\alpha)) \; f_A(\alpha; a, \bday | h) d\alpha}
}
