% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nzPois}
\alias{nzPois}
\title{Compute N non-zero values from a Poisson distribution with a given MoI}
\usage{
nzPois(N, moi)
}
\arguments{
\item{N}{the number of random variates to return}

\item{moi}{the mean moi}
}
\value{
a \link{numeric} vector of length N
}
\description{
Compute N non-zero values from a Poisson distribution with a given MoI
}
