% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pois.R
\name{p_counts.pois}
\alias{p_counts.pois}
\title{Detection of infection given parasitemia}
\usage{
\method{p_counts}{pois}(x, xi, bvm = 13, pars = par_pois())
}
\arguments{
\item{x}{raw parasite counts}

\item{xi}{\eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{pars}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
