% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-interface.R
\name{p_nz_counts_log_binned}
\alias{p_nz_counts_log_binned}
\title{The binned PDF for non-zero counts, \eqn{\log_{10}} transformed}
\usage{
p_nz_counts_log_binned(xi, bins = NULL, bvm = 13, pars = par_nb())
}
\arguments{
\item{xi}{\eqn{\log_{10}} parasite densities}

\item{bins}{breakpoints for summarizing outputs as \eqn{\log_{10}} counts}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{pars}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
The binned PDF for non-zero counts, \eqn{\log_{10}} transformed
}
