% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-interface.R
\name{p_Omega}
\alias{p_Omega}
\title{The density function for parasite densities as a function of the mean}
\usage{
p_Omega(xi, mu, bvm = 13, par_Omega = par_Omega_beta())
}
\arguments{
\item{xi}{a vector of probabilities for \eqn{log_{10}} parasite densities}

\item{mu}{the expected value for \eqn{log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{log_{10}} red blood cells}

\item{par_Omega}{parameters defining parasite densities as a function of the mu}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The density function for parasite densities as a function of the mean
}
