% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{pAoI}
\alias{pAoI}
\title{Compute the distribution function for AoI}
\usage{
pAoI(alpha, a, FoI_a, bday = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{alpha}{the age of an infection, \eqn{\alpha}}

\item{a}{cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
The distribution function for the AoI is given by
\deqn{F_A(\alpha | a, \bday, h) = \int_0^\alpha f_A(\alpha, a, \bday | h) d \alpha}
}
