% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoy.R
\name{pAoY}
\alias{pAoY}
\title{The distribution function for the age of the youngest infection (AoY)}
\usage{
pAoY(alpha, a, FoI_a, bday = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{alpha}{the age of an infection, \eqn{\alpha}}

\item{a}{cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
The distribution function for the age of the youngest infection (AoY)
}
