% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-seasonal.R
\name{par_expSeason}
\alias{par_expSeason}
\title{Return a list to dispatch seasonalFoI.exp}
\usage{
par_expSeason(phase = 0, pwr = 1, lift = 0, shift = 0)
}
\arguments{
\item{phase}{shift the phase / time of year}

\item{pwr}{raise the seasonal pattern to a power}

\item{lift}{the amplitude of the}

\item{shift}{shift everything up or down}
}
\value{
a \link{list}
}
\description{
Return a list to dispatch seasonalFoI.exp
}
