% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-seasonal.R
\name{par_flatSeason}
\alias{par_flatSeason}
\title{Return a list that dispatches seasonalFoI.flat}
\usage{
par_flatSeason()
}
\value{
a \link{list}
}
\description{
Return a list that dispatches seasonalFoI.flat
}
