% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmu-base.R
\name{par_Fmu_base}
\alias{par_Fmu_base}
\title{Set up parameters for \link{Fmu.base}}
\usage{
par_Fmu_base(peak = 20, liver = 7, tildeb = 10.3, tildel = 2, Sa = 0.0033)
}
\arguments{
\item{peak}{The age of infection (in days) when parasite densities peak}

\item{liver}{The age of infection (in days) when parasites emerge from the liver}

\item{tildeb}{The maximum expected log10 parasite densities}

\item{tildel}{The minimum expected log10 parasite densities}

\item{Sa}{The decline in mu with respect to alpha}
}
\value{
a \link{list}
}
\description{
Set up parameters for \link{Fmu.base}
}
