% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBC.R
\name{par_lRBC_static}
\alias{par_lRBC_static}
\title{Set up parameters for [log10RBC.static]}
\usage{
par_lRBC_static(lRBCmax = 13)
}
\arguments{
\item{lRBCmax}{the maximum}
}
\value{
a [list]
}
\description{
Set up parameters for [log10RBC.static]
}
