% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-nb.R
\name{par_nb}
\alias{par_nb}
\title{Parameters for negative binomial sampling}
\usage{
par_nb(sz = 0.31, s = 6)
}
\arguments{
\item{sz}{the negative binomial size parameter}

\item{s}{sample volume as log10 RBCs}
}
\value{
binary detection result
}
\description{
Parameters for negative binomial sampling
}
