% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-pois.R
\name{par_pois}
\alias{par_pois}
\title{Parameters for Poisson sampling}
\usage{
par_pois(s = 6)
}
\arguments{
\item{s}{sample volume as log10 RBCs}
}
\value{
par a \link{list}
}
\description{
Parameters for Poisson sampling
}
