% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-beta.R
\name{par_sigma_abc}
\alias{par_sigma_abc}
\title{Parameters to configure \link{sigma_mu.abc}}
\usage{
par_sigma_abc(aa = 3.11, bb = 2.14, cc = 1.74)
}
\arguments{
\item{aa}{a shape parameter}

\item{bb}{a shape parameter}

\item{cc}{a shape parameter}
}
\value{
a \link{list}
}
\description{
Parameters to configure \link{sigma_mu.abc}
}
