% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-age.R
\name{par_type2Age}
\alias{par_type2Age}
\title{Make a parameter list to dispatch ageFoI.type2}
\usage{
par_type2Age(shift = 0.1, A = 1.5, B = 2)
}
\arguments{
\item{shift}{a shape parameter}

\item{A}{a shape parameter}

\item{B}{the age weight for adults}
}
\value{
\link{list}
}
\description{
Make a parameter list to dispatch ageFoI.type2
}
