% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-beta.R
\name{pbeta1}
\alias{pbeta1}
\title{Disribution function for the beta distribution, an alternative parameterization}
\usage{
pbeta1(p, mu, pSig = par_sigma_abc())
}
\arguments{
\item{p}{a vector of probabilities}

\item{mu}{the mean value for the distribution (0 <= mu <= 1)}

\item{pSig}{parameters to dispatch the S3 function \link{sigma_mu}}
}
\value{
a \link{numeric} vector of length(p)
}
\description{
Disribution function for the beta distribution, an alternative parameterization
}
