% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoyN.R
\name{pYoN}
\alias{pYoN}
\title{The youngest of N infections, distribution function}
\usage{
pYoN(N, a, FoI_a, bday = 0, hhat = NULL, r = 1/200)
}
\arguments{
\item{N}{the MoI}

\item{a}{the age of a cohort}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate of a simple infection}
}
\value{
a \link{numeric} vector of length a + 1
}
\description{
The youngest of N infections, distribution function
}
