% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2density.R
\name{r_alpha2density}
\alias{r_alpha2density}
\title{The distribution function for parasite densities in a simple malaria infection of age alpha}
\usage{
r_alpha2density(
  n,
  alpha,
  W = 0,
  a = 0,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_Omega = par_Omega_beta()
)
}
\arguments{
\item{n}{the number of observations}

\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{a}{host cohort age}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{Fmu}}

\item{par_Omega}{parameters to compute \link{r_Omega}}
}
\value{
a \link{numeric} vector of length(n)
}
\description{
The distribution function for parasite densities in a simple malaria infection of age alpha
}
