% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-beta.R
\name{r_Omega.beta}
\alias{r_Omega.beta}
\title{Modified beta distribution, random numbers}
\usage{
\method{r_Omega}{beta}(n, mu, bvm = 13, par_Omega = par_Omega_beta())
}
\arguments{
\item{n}{the number of observations}

\item{mu}{the expected value for \eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{par_Omega}{parameters defining parasite densities as a function of the mu}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The beta distribution, parameterized by the mean and variance, modified to return
a number between 0 and bvm (the number of red blood cells)
}
