% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-interface.R
\name{r_Omega}
\alias{r_Omega}
\title{Random generation of parasite densities from a simple malaria infection}
\usage{
r_Omega(n, mu, bvm = 13, par_Omega = par_Omega_beta())
}
\arguments{
\item{n}{the number of observations}

\item{mu}{the expected value for \eqn{\log_{10}} parasite densities}

\item{bvm}{blood volume as \eqn{\log_{10}} red blood cells}

\item{par_Omega}{parameters defining parasite densities as a function of the mu}
}
\value{
a \link{numeric} vector of length(n)
}
\description{
Random generation of parasite densities from a simple malaria infection
}
