% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{rAoI}
\alias{rAoI}
\title{Random numbers for the AoI}
\usage{
rAoI(N, a, FoI_a, bday = 0, hhat = 1, r = 1/200, alphamin = 0)
}
\arguments{
\item{N}{the number of observations}

\item{a}{cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{alphamin}{the minimum value of the AoI to return}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
Draw random numbers for the AoI from a cohort, \eqn{\hat A_\bday(a)}
}
