% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-beta.R
\name{rbeta1}
\alias{rbeta1}
\title{The random generation function for the beta distribution, an alternative parameterization
Title}
\usage{
rbeta1(n, mu, pSig = par_sigma_abc())
}
\arguments{
\item{n}{number of observations}

\item{mu}{the mean value for the distribution (0 <= mu <= 1)}

\item{pSig}{parameters to dispatch the S3 function \link{sigma_mu}}
}
\value{
a \link{numeric} vector of length n
}
\description{
The random generation function for the beta distribution, an alternative parameterization
Title
}
