% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_parasites.R
\name{rRda}
\alias{rRda}
\title{Random generation for M parasite densities in a host cohort with MoI}
\usage{
rRda(
  M,
  R,
  a,
  FoI_a,
  bday = 0,
  hhat = 1,
  r = 1/200,
  alphamin = 7,
  RBC_par = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  Omega_par = par_Omega_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{M}{the MoI}

\item{R}{the number of observations}

\item{a}{host cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{alphamin}{the minimum value of alpha allowed}

\item{RBC_par}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{Fmu}}

\item{Omega_par}{parameters defining parasite densities as a function of the mu}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a R by M \link{matrix}
}
\description{
Random generation for M parasite densities in a host cohort with MoI
}
