% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoyN.R
\name{rYoN}
\alias{rYoN}
\title{The youngest of N infections, random numbers}
\usage{
rYoN(R, N, a, FoI_a, bday = 0, hhat = NULL, r = 1/200, alphamin = 0)
}
\arguments{
\item{R}{the number of observations}

\item{N}{the number of infections per person}

\item{a}{the host cohort age}

\item{FoI_a}{parameters that define an FoI function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{alphamin}{the minimum value of the AoI to return}
}
\value{
a \link{numeric} vector of length R
}
\description{
The youngest of N infections, random numbers
}
