% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-seasonal.R
\name{seasonalFoI.exp}
\alias{seasonalFoI.exp}
\title{A generalized sinusoidal seasonal pattern, exponentiated}
\usage{
\method{seasonalFoI}{exp}(t, par = par_expSeason())
}
\arguments{
\item{t}{the time}

\item{par}{a list}
}
\value{
\link{numeric}
}
\description{
A generalized sinusoidal seasonal pattern, exponentiated
}
