% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-seasonal.R
\name{seasonalFoI.flat}
\alias{seasonalFoI.flat}
\title{The "no seasonality" function}
\usage{
\method{seasonalFoI}{flat}(t, par)
}
\arguments{
\item{t}{the time}

\item{par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
The "no seasonality" function
}
