% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-seasonal.R
\name{seasonalFoI}
\alias{seasonalFoI}
\title{Add a seasonal pattern to the FoI trace function}
\usage{
seasonalFoI(t, par)
}
\arguments{
\item{t}{the time}

\item{par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Add a seasonal pattern to the FoI trace function
}
