% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-seasonal.R
\name{seasonalFoI.sin}
\alias{seasonalFoI.sin}
\title{A generalized sinusoidal seasonal pattern}
\usage{
\method{seasonalFoI}{sin}(t, par = par_sinSeason())
}
\arguments{
\item{t}{the time}

\item{par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
A generalized sinusoidal seasonal pattern
}
