% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmu-interface.R
\name{sFmu}
\alias{sFmu}
\title{Compute mean, expected parasite densities \code{mu} for multiple values of the age of infection \code{alpha}}
\usage{
sFmu(alpha, W, par_Fmu)
}
\arguments{
\item{alpha}{a parasite's age of infection}

\item{W}{immune tracking variables}

\item{par_Fmu}{a \link{list} that defines a model}
}
\value{
mean log10 parasite densities, a \link{numeric} vector of length(alpha)
}
\description{
Compute mean, expected parasite densities \code{mu} for multiple values of the age of infection \code{alpha}
}
