% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-beta.R
\name{sigma_mu.abc}
\alias{sigma_mu.abc}
\title{A function that returns constrained values of the variance for the beta distrution as a function of the mean.}
\usage{
\method{sigma_mu}{abc}(mu, par)
}
\arguments{
\item{mu}{the mean value for the distribution (0 <= mu <= 1)}

\item{par}{parameters to dispatch and configure the instances}
}
\value{
a \link{numeric} vector of length(mu)
}
\description{
A function that returns constrained values of the variance for the beta distrution as a function of the mean.
}
