% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega-beta.R
\name{sigma_mu}
\alias{sigma_mu}
\title{A function to compute the variance of the beta distrution as a function of the mean.}
\usage{
sigma_mu(mu, par)
}
\arguments{
\item{mu}{the mean value for the distribution (0 <= mu <= 1)}

\item{par}{parameters to dispatch and configure the instances}
}
\value{
a \link{numeric} vector of length(mu)
}
\description{
A function to compute the variance of the beta distrution as a function of the mean.
}
