% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{solve_dm}
\alias{solve_dm}
\title{Solve the hybrid model for the MoI}
\usage{
solve_dm(h, FoI_a, r = 1/200, bday = 0, Amax = 730, dt = 1)
}
\arguments{
\item{h}{the average, annual force of infection}

\item{FoI_a}{a cohort trace function}

\item{r}{the clearance rate for a simple infection}

\item{bday}{the cohort birthday}

\item{Amax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}
}
\value{
a \link{matrix} with the orbits
}
\description{
Solve the hybrid model for the MoI
}
