% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{solve_dpda}
\alias{solve_dpda}
\title{Solve a system of differential equations to compute the true PR and the MoI}
\usage{
solve_dpda(h, FoI_a, r = 1/200, bday = 0, Amax = 730, dt = 1)
}
\arguments{
\item{h}{a scaling factor on the FoI}

\item{FoI_a}{a cohort trace function}

\item{r}{the clearance rate for a simple infection}

\item{bday}{the cohort birthday}

\item{Amax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}
}
\value{
a \link{data.frame} with the orbits
}
\description{
Solve a system of differential equations to compute the true PR and the MoI
}
