% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquIP.R
\name{SquIP}
\alias{SquIP}
\title{SquIP}
\description{
A model for the
distribution of the multiplicity of infection (MoI) in a cohort of humans
as it ages.
The model extends a basic queuing model that has been used in malaria epidemiology before called \eqn{M/M/\infty}: it adds parameters
to model treatment with anti-malarial drugs, which cures infections and is followed
by a short period of chemo-protection. The model is named after its states:
}
\section{\strong{SquIP}  \eqn{=}}{


\strong{S} - \strong{S}usceptible \eqn{+}

\strong{quI} - a \strong{qu}euing process for \strong{I}nfected \eqn{+}

\strong{P} - chemo-\strong{P}rotected

The master equations are presented below. The software implementation of \strong{SquIP} has two parts:
\itemize{
\item \link{dSquIP} computes the derivatives
\item \link{solve_SquIP} is a wrapper around \link{dSquIP} that
\itemize{
\item sets the value of a truncation parameter
\item sets up the initial values, times, and parameters
\item calls \link[deSolve:ode]{deSolve::ode} to solve the system of ordinary differential equations
\item parses the outputs
}
}

This software implementation truncates the infinite system of dIifferential equations. To
check the accuracy of the truncated system, we have derived hybrid variables.
}

\section{The Dynamical System}{


\strong{Variables}

The independent variable is age, \eqn{a.}

The dependent variables are:
\itemize{
\item \eqn{I_i} is the expected number infected with \eqn{i} clones
\item \eqn{S} is the expected number uninfected and susceptible to infection \eqn{(S=I_0)}
\item \eqn{P} is the expected number uninfected and chemo-protected
\item \eqn{H = S+P+\sum_i I_i} is expected number in the cohort
}

The model assumes everyone is born susceptible
so \eqn{H} is passed as a parameter, and the initial conditions are set to \eqn{S(0) = H} and \eqn{I_i(0) = P(0)=0.}

\strong{Exposure} --- \eqn{h(a)}

Let \eqn{h(a)} denote the force of infection (FoI). This model assumes
that each incident infection increases the MoI by exactly one. The FoI is passed as a trace function.

\strong{Natural Clearance} --- \eqn{r}

The model assumes that each clone clears independently of the others at rate \eqn{r,} so if a person has \eqn{i} clones, the MoI goes down by one at the rate \eqn{ri.}

\strong{Treatment and Chemoprotection} --- \eqn{\xi, \rho, \sigma, \eta}

These equations assume that individuals are treated and cured for several different reasons:
\itemize{
\item Everyone in the population who is not chemoprotected takes drugs at a background rate \eqn{\xi}
\item Some incident infections cause disease, and a fraction \eqn{\rho} gets treated
\item In addition to these other modes of treatment, infected individuals get treated at the higher rate \eqn{\sigma}
}

After being treated and cured, individuals enter the chemoprotected class \eqn{P.} Chemoprotection is lost
at the rate \eqn{\eta,} and they enter the susceptible class \eqn{S.}

\strong{Demography} --- \eqn{\mu}

The model assumes that all individuals die at the rate \eqn{\mu(a)}, such that

\deqn{\frac{\textstyle{dH}}{\textstyle{da}} = -\mu H}
}

\section{Master Equations}{


The dynamics are described by an infinite set of differential equations:
\deqn{
\begin{array}{rl}
 dS/da  &= \eta P + r I_1 - (h + \xi + \mu) S \\
 dP/da  &=  \left(h \rho + \xi \right) \left(H-P \right) + \sigma \sum_i I_i   - \left(\eta + \mu \right) P \\
 dI_1/ da  &=  h(1-\rho)S + 2 r I_2 - (r + h + \xi + \sigma + \mu) I_1 \\
 dI_i/da  &=  h(1-\rho)I_{i-1} + (i+1) r I_{i+1} - (i r + h + \xi + \sigma + \mu) I_i\\
\end{array}
}

In this software implementation, the infinite system of equations is truncated after \eqn{i=N} and
\deqn{dI_N/da  =  h(1-\rho)I_{N-1} - (N r + h + \xi + \sigma + \mu) I_N}
}

\section{Hybrid Variables}{


To check the accuracy of the truncated system, we have derived hybrid variables for the first
two moments of the distribution of the MoI.

Let \eqn{m_n} denote the \eqn{n^{th}} moment of the MoI:
\deqn{m_n = \frac{\textstyle{1}}{\textstyle{H}}\sum_i i^n I_i}

This software implementation computes the dynamics of the hybrid variables \eqn{m_1} and \eqn{m_2}:
\deqn{
\begin{array}{rl}
d m_1 /da &=  h(1-\rho)\left(1-P/H\right) - \left(r + h \rho +\sigma + \xi\right) m_1 \\
d m_2 /da &=  h(1-\rho)\left( 1- P/H + 2 m_1\right)  - \left(2r + h \rho + \sigma + \xi \right) m_2 + r m_1 \\
\end{array}
}

After solving, the output is parsed and the
first two moments of the distribution of the MoI are computed from \eqn{I_i}.
The empirical moments should match \eqn{m_1} and \eqn{m_2} unless the value of the
truncation parameter \eqn{N} is set too low.
}

\seealso{
\link{dSquIP} & \link{solve_SquIP} |  The model \link{SquIPz} extends \strong{SquIP.} | \link{SIPm} is an approximating model.
}
