% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquIPz.R
\name{SquIPz}
\alias{SquIPz}
\title{SquIPz}
\description{
This is a model for the
dIistribution of the multiplicity of infection (MoI) in a cohort of humans
as it ages. It extends the queuing model \link{SquIP} by adding a Tweedie process to model
heterogeneous exposure: each infection event increases the MoI by the multiplicity of exposure (MoE), which is given by
a probability mass function
\eqn{F_\mbox{MoE}.} The model is named after its states and processes:
}
\section{\strong{SquIPz}  \eqn{=}}{


\strong{S} - \strong{S}usceptible \eqn{+}

\strong{quI} - a \strong{qu}euing process for \strong{I}nfected \eqn{+}

\strong{P} - chemo-\strong{P}rotected

\strong{z} - the Tweedie process distribution

This software implementation of \strong{SquIP} has two parts:
\itemize{
\item \link{dSquIPz} computes the derivatives
\item \link{solve_SquIPz} is a wrapper around \link{dSquIPz} that
\itemize{
\item sets the value of a truncation parameter
\item sets up the initial values, times, and parameters
\item calls \link[deSolve:ode]{deSolve::ode} to solve the system of ordinary differential equations
\item parses the outputs
}
}

This software implementation truncates the infinite system of dIifferential equations. To
check the accuracy of the truncated system, we have derived hybrid variables.
}

\section{The Model}{


\strong{Variables} ---
The independent variable is age, \eqn{a.}

The dependent variables are:
\itemize{
\item \eqn{I_i} is the expected number infected with \eqn{i} clones
\item \eqn{S} is the expected number uninfected and susceptible to infection \eqn{(S=I_0)}
\item \eqn{P} is the expected number uninfected and chemo-protected
\item \eqn{H = S+P+\sum_i I_i} is expected number in the cohort
}

The model assumes everyone is born susceptible
so \eqn{H} is passed as a parameter, and the initial conditions are set to \eqn{S(0) = H} and \eqn{I_i(0) = P(0)=0.}

\strong{Exposure} (\eqn{h}) ---

Let \eqn{h} denote the force of infection (FoI). This model assumes
that each incident infection increases the MoI by \eqn{i} with probability \deqn{F_\mbox{MoE}(X=i) = z_i.}

\strong{Natural Clearance} ---
The model assumes that each clone clears independently of the others at rate \eqn{r,} so if a person has \eqn{i} clones, the MoI goes down by one at the rate \eqn{ri.}

\strong{Treatment and Chemoprotection} ---
These equations assume that individuals are treated and cured for several different reasons:
\itemize{
\item Everyone in the population takes drugs at a background rate \eqn{\xi}
\item Incident infections cause disease and a fraction \eqn{\rho} gets treated
\item In addition to other modes of treatment, infected individuals get treated at the higher rate \eqn{\sigma}
}

After being treated and cured, individuals enter the chemoprotected class \eqn{P.} Chemoprotection is lost
at the rate \eqn{\eta,} and they enter the susceptible class \eqn{S.}

\strong{Demography} --
The model assumes that all individuals die at the rate \eqn{\mu(a)}, such that

\deqn{\frac{\textstyle{dH}}{\textstyle{da}} = -\mu H}

\strong{Dynamics} ---
The dynamics are described by an infinite set of differential equations:
\deqn{
\begin{array}{rl}
 dS/da  &= \eta P + r I_1 - (h + \xi + \mu) S \\
 dP/da  &=  \left(h \rho + \xi \right) \left(H-P \right) + \sigma \sum_i I_i   - \left(\eta + \mu \right) P \\
 dI_i / da &=h (1-\rho) \left( S z_i + \sum_j  I_{i-j} z_j \right)  + (i+1) r I_{i+1} - \left(h + \sigma + r i + \xi + \mu \right) I_i\\
\end{array}
}

In this software implementation, the infinite system of equations is truncated at \eqn{N} and
\deqn{dI_N/da  =  h (1-\rho) \sum_{i \geq N} \left( S  z_i + \sum_j  I_{i-j} z_j \right)  - \left(h + \sigma + r N + \xi + \mu \right) I_N}
}

\section{Hybrid Variables}{


Let \eqn{m_n} denote the  \eqn{n^{th}} moment of the MoI:
\deqn{m_n = \frac{\textstyle{1}}{\textstyle{H}}\sum_i i^n I_i}
and let \eqn{\tilde z_n} denote the moments of
\eqn{F_\mbox{MoE}:}
\deqn{\tilde z_n = \sum_i i^n z_i}

This software implementation computes the dynamics of the hybrid variables \eqn{m_1} and \eqn{m_2}:
\deqn{
\begin{array}{rl}
d m_1 /da &=  h(1-\rho)\tilde z_1 \left(1-P/H \right) - \left(r + h \rho +\sigma + \xi\right) m_1 \\
d m_2 /da &=   h(1-\rho)\hat z_2 \left( 1-P/H\right)  + (r+2 h(1-\rho)\hat z) m  - \left(2r + h \rho + \sigma + \xi \right) m_2 \\
\end{array}
}
After solving, the output is parsed and the
first two moments of the distribution of the MoI are computed from \eqn{I_i}.
The empirical moments should match \eqn{m_1} and \eqn{m_2} unless the value of the
truncation parameter \eqn{N} is set too low.
}

\seealso{
\link{SquIP} and \link{SIPm}
}
