% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-trend.R
\name{trendFoI.flat}
\alias{trendFoI.flat}
\title{The "no trend" function}
\usage{
\method{trendFoI}{flat}(t, par)
}
\arguments{
\item{t}{the time}

\item{par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
The "no trend" function
}
