% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi-trend.R
\name{trendFoI}
\alias{trendFoI}
\title{Add a trend to the FoI trace function}
\usage{
trendFoI(t, par)
}
\arguments{
\item{t}{the time}

\item{par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Add a trend to the FoI trace function
}
