% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_dynamics.R
\name{truePR}
\alias{truePR}
\title{Compute the true PR in a cohort as a function of age and exposure}
\usage{
truePR(a, FoI_a, bday = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{a}{the host cohort age}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
The true PR is \deqn{p_\bday(a|h) = 1 - e^{-m_\bday(a|h)}}
}
