% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_tracking.R
\name{Wda}
\alias{Wda}
\title{Compute immune tracking variables as a function of host age and exposure}
\usage{
Wda(a, FoI_a, bday = 0, hhat = 1, par = par_Wda_none())
}
\arguments{
\item{a}{age of a host cohort}

\item{FoI_a}{a cohort trace function}

\item{bday}{cohort birthday}

\item{hhat}{overrides the value of hbar in par}

\item{par}{parameters in a \link{list}}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
The function dispatches on \code{class(par)}
}
