% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_dynamics.R
\name{zda}
\alias{zda}
\title{Compute infection density in a cohort of humans, \eqn{z_\bday(\alpha, a |h)}}
\usage{
zda(alpha, a, FoI_a, bday = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{alpha}{the age of an infection, \eqn{\alpha}}

\item{a}{the age of the cohort}

\item{FoI_a}{a cohort trace function}

\item{bday}{the cohort birthday}

\item{hhat}{scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
Given a function describing the FoI (\eqn{h(a,d)}), and a parameter
describing the clearance rate of infections (\eqn{r}),
the density of parasite clones of age \eqn{\alpha} distributed among a cohort of humans of
age \eqn{a} is \deqn{z_d(\alpha, a) = e^{-r \alpha} h(a-\alpha, d)}
}
