% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_parasites.R, R/detection.R
\name{d_moi_count}
\alias{d_moi_count}
\title{Detection of infection given parasitemia}
\usage{
d_moi_count(
  a,
  FoIpar,
  tau = 0,
  hhat = 1,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  par_Omega = par_Omega_beta(),
  pWda = par_Wda_none(),
  par_sample = par_nb()
)

d_moi_count(
  a,
  FoIpar,
  tau = 0,
  hhat = 1,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  par_Omega = par_Omega_beta(),
  pWda = par_Wda_none(),
  par_sample = par_nb()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{tau}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{Fmu}}

\item{par_Omega}{parameters to compute \link{d_Omega}}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{par_sample}{parameters that define a detection function}
}
\value{
detection probability

detection probability
}
\description{
Detection of infection given parasitemia

Detection of infection given parasitemia
}
