% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{dAoIda}
\alias{dAoIda}
\title{Compute the derivatives of the AoI moments dynamically}
\usage{
dAoIda(a, M, p, FoIpar)
}
\arguments{
\item{a}{the host age}

\item{M}{the state variables}

\item{p}{the parameters}

\item{FoIpar}{parameters that define an FoI function}
}
\value{
the derivatives of the MoI and the AoI as a \link{list}
}
\description{
The
}
