% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquIP.R
\name{dSquIP}
\alias{dSquIP}
\title{Derivatives for \link{SquIP}}
\usage{
dSquIP(a, y, pars, F_a)
}
\arguments{
\item{a}{the host age}

\item{y}{a vector of state variables}

\item{pars}{the parameters}

\item{F_a}{a trace function}
}
\value{
the derivatives as a \link{list}
}
\description{
This function computes the derivatives for the queuing model \link{SquIP},
in a form that can be used by \link[deSolve:ode]{deSolve::ode}.

It also computes the derivatives for the hybrid variables \eqn{m_1} and \eqn{m_2}
}
\seealso{
\link{SquIP} & \link{solve_SquIP}

\link{solveMMinfty}
}
\keyword{internal}
