% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmu-interface.R
\name{Fmu}
\alias{Fmu}
\title{Compute mean, expected parasite densities \code{mu} as a function of the age of infection \code{alpha}}
\usage{
Fmu(alpha, W, par_Fmu)
}
\arguments{
\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{par_Fmu}{a \link{list} that defines a model}
}
\value{
mean log10 parasite densities, a \link{numeric} vector of length(alpha)
}
\description{
Compute mean, expected parasite densities \code{mu} as a function of the age of infection \code{alpha}
}
