% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{FQ}
\alias{FQ}
\title{The proportion of zero counts in}
\usage{
FQ(
  alpha,
  a,
  FoIpar,
  tau = 0,
  hhat = 1,
  r = 1/200,
  svm = 1e+06,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_Omega = par_Omega_beta(),
  pWda = par_Wda_none(),
  par_sample = par_nb()
)
}
\arguments{
\item{alpha}{the age of a parasite infection}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{tau}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{svm}{the volume of a blood sample}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{Fmu}}

\item{par_Omega}{parameters to compute \link{d_Omega}}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{par_sample}{parameters that define a detection function}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The proportion of zero counts in
}
