% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquIPz.R
\name{make_Z}
\alias{make_Z}
\title{Make the Z matrix}
\usage{
make_Z(d_pmf, N)
}
\arguments{
\item{N}{the maximum MoI}

\item{F_moe}{the density distribution for a probability mass function}
}
\value{
a \link{matrix}
}
\description{
Make the Z matrix
}
\examples{
dd = function(x){dpois(x,.2)}
make_Z(dd, 20)
}
\keyword{internal}
