% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{mean_clone_counts}
\alias{mean_clone_counts}
\title{Compute the mean parasite counts in simple infections}
\usage{
mean_clone_counts(
  a,
  FoIpar,
  dx = 0.1,
  hhat = 1,
  tau = 0,
  r = 1/200,
  Fmu_par = par_Fmu_base(),
  Omega_par = par_Omega_beta(),
  pWda = par_Wda_none(),
  par_sample = par_nb()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{dx}{the width of the mesh for computing the CDF of parasite densities}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{Fmu_par}{parameters to compute \link{Fmu}}

\item{Omega_par}{parameters defining parasite densities as a function of the mu}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{par_sample}{parameters that define a detection function}
}
\value{
a \link{list}
}
\description{
Compute the mean parasite counts in simple infections
}
