% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_dynamics.R
\name{meanMoI}
\alias{meanMoI}
\title{The mean MoI in a host cohort of age \eqn{a}}
\usage{
meanMoI(a, FoIpar, tau = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{a}{host age}

\item{FoIpar}{a compound \link{list} to compute \eqn{h_\tau(a)}}

\item{tau}{the host cohort's birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
The mean multiplicity of infection (MoI) is \deqn{m_\tau(a|h) = \int_0^a z_\tau(\alpha, a|h) d \alpha}
}
