% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMinfinity.R
\name{MoIDistPlot}
\alias{MoIDistPlot}
\title{Plot the output of \link{solveMMinfty}}
\usage{
MoIDistPlot(moi, t, clr1 = "red")
}
\arguments{
\item{moi}{the mean moi}

\item{t}{the time}

\item{clr1}{the color}
}
\description{
Plot the output of \link{solveMMinfty}
}
