% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{momentAoI}
\alias{momentAoI}
\title{Compute the moments for the AoI density function for a cohort of age a}
\usage{
momentAoI(a, FoIpar, tau = 0, hhat = 1, r = 1/200, n = 1)
}
\arguments{
\item{a}{host age}

\item{FoIpar}{a compound \link{list} to compute \eqn{h_\tau(a)}}

\item{tau}{the host cohort's birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{n}{the moment desired}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
Compute the moments for the AoI density function for a cohort of age a
}
