% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_parasites.R
\name{p_parasite_density}
\alias{p_parasite_density}
\title{Call \link{parasite_density} and return the density vector}
\usage{
p_parasite_density(
  meshX,
  a,
  FoIpar,
  tau = 0,
  hhat = 1,
  r = 1/200,
  RBC_par = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  Omega_par = par_Omega_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{meshX}{a mesh over parasite densities}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{tau}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{RBC_par}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{Fmu}}

\item{Omega_par}{parameters to compute \link{d_Omega}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a numeric vector of length meshX
}
\description{
Call \link{parasite_density} and return the density vector
}
