% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoy.R
\name{pAoY_long}
\alias{pAoY_long}
\title{Alternative method for computing the distribution function for the age of the youngest infection (AoY)}
\usage{
pAoY_long(alpha, a, FoIpar, tau = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{alpha}{the age of an infection, \eqn{\alpha}}

\item{a}{host age}

\item{FoIpar}{a compound \link{list} to compute \eqn{h_\tau(a)}}

\item{tau}{the host cohort's birthday}

\item{hhat}{scaling parameter for \link{FoI}}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
This method computes the distribution function for the AoY by summing over AoI distributions
}
