% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquIP.R
\name{parse_SquIP}
\alias{parse_SquIP}
\title{Parse SquIP}
\usage{
parse_SquIP(pars, out)
}
\arguments{
\item{pars}{the parameters, a named list}

\item{out}{the orbits}
}
\value{
a named list
}
\description{
Parse SquIP
}
\keyword{internal}
