% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPm.R
\name{SIPm}
\alias{SIPm}
\title{SIPm}
\description{
\strong{SIPm} is an enhanced compartmental model that approximates the model \strong{SquIPz}
}
\section{Variables}{


The independent variable is age, \eqn{a.}

Three variables make up a compartmental model:
\itemize{
\item \eqn{S} is the expected number uninfected and susceptible to infection
\item \eqn{P} is the expected number uninfected and chemo-protected
\item \eqn{I} is the expected number infected
}

Two variables track the distribution of the MoI
\itemize{
\item \eqn{m_1} is the first moment of the MoI
\item \eqn{m_2} is the second moment of the MoI
}
}

\section{Demography}{


We let \deqn{H = S+I+P} denote the expected population density of the cohort as it ages.
The model assumes that all individuals die at the rate \eqn{\mu}, such that

\deqn{\frac{\textstyle{dH}}{\textstyle{da}} = -\mu H}

In the model, we assume that everyone is born susceptible
so \eqn{H} is passed as a parameter, and the initial conditions are set to \eqn{S(0) = H} and \eqn{I_i(0) = P(0)=0.}
}

\section{Exposure}{

We let \eqn{h} denote the force of infection (FoI), and we let \eqn{F_\mbox{MoE}} denote the distribution of the multiplicity of exposure (MoE).
\itemize{
\item \eqn{\tilde z_1} is the first moment of \eqn{F_\mbox{MoE}}
\item \eqn{\tilde z_2} is the second moment of \eqn{F_\mbox{MoE}}
}
}

\section{Clearance}{

In the queuing model \link{SquIP}, infections clear at the rate \eqn{r I_1.} If we think of \strong{SIPm} as an approximating model, then
\deqn{r I_1 \approx  F_r(m_1, m_2) I}

To get $F_r,$ we assume that the distribution of the MoI in these models is
approximately a zero-inflated negatively binomial distributed.
We assume
\deqn{
\frac{I_i} H\sim \mbox{NB}\left(\zeta=i| m, \phi \right)\left(1-\frac PH\right)
}
where \verb{mu1=\eqn{m_1} is the mean and }size`\eqn{=\phi} (the size parameter)  is given by:
\deqn{\phi = \frac{m_1^2}{m_2-m_1^2-m_1}}
so
\deqn{
F_r(m, \phi) = \begin{cases}
 r & \mbox{if } m = 0 \\
 r \frac{
 \mbox{NB}\left(\zeta=1| m, \phi \right)
 }{1-
 \mbox{NB}\left(\zeta=0 | m, \phi \right)
 } \left(1 - \frac{P}{H} \right) & \mbox{if } m>0
 \end{cases}
 }

In effect, the approximation assumes \eqn{I_1/H} is close to the probability the MoI is equal to 1 in a zero-inflated negative binomial distribution. The distributions of the MoI are probably not exactly negatively binomially distributed. We use the Kullback-Liebler divergence to evaluate how well a negative binomial approximates the distributions in the queueing models.
}

\section{Treatment and Chemprotection}{


These equations assume that individuals are treated and cured for several different reasons:
\itemize{
\item Incident infections cause disease and a fraction \eqn{\rho} gets treated
\item Everyone in the population takes drugs at a background rate \eqn{\xi}
\item In addition to other modes of treatment, infected individuals get treated at the higher rate \eqn{\sigma}
}

After being treated and cured, individuals enter the chemoprotected class \eqn{P.} Chemoprotection is lost
at the rate \eqn{\eta,} and they enter the susceptible class \eqn{S.}
}

\section{Differential Equations}{


\strong{Hybrid Variables}
The compartmental dynamics are enhanced by computing a hybrid variable, a non-compartmental state variable describing the mean MoI (\eqn{m_1}), with dynamics:
\deqn{
\frac{\textstyle{dm_1}}{\textstyle{da}} =  h (1-\rho) \hat z \left( 1-\frac{\textstyle{P}}{\textstyle{H}} \right) - \left(r + h \rho + \sigma + \xi \right) m_1}

We also compute the dynamics of the second moment of the distribution of the MoI:
\deqn{
\begin{array}{rl}
\frac{\textstyle{dm_2}}{\textstyle{da}}  = & h(1-\rho)\hat z_2 \left( 1- \frac{  \textstyle{P}}{\textstyle{H}}\right)  + (r+2 h(1-\rho)\hat z) m_1  \\[6pt]
&- \left(2r + h \rho + \sigma + \xi \right) m_2  \\
\end{array}}

\strong{Compartmental Variables}

The infinite infection states in the family of queuing models specified by \link{SquIPz} are compressed down to a finite set of mutually exclusive and collectively exhaustive states -- \eqn{S}, \eqn{I}, and \eqn{P} -- with dynamics:
\deqn{
\begin{array}{rl}
\frac{\textstyle{dS}}{\textstyle{da}} &= F_r(m, \phi) I + \eta P - (h + \xi + \mu) S \\[6pt]
\frac{\textstyle{dI}}{\textstyle{da}} &= h \left(1-\rho\right) S  \\[6pt] & - \left( \rho h + \xi + \sigma + F_r(m, \phi) + \mu  \right) I \\[6pt]
\frac{\textstyle{dP}}{\textstyle{da}} &= (h \rho+ \xi)(H-P)  + \sigma I - (\eta+\mu) P \\[6pt]
\end{array}}
}

\section{Parameters}{

}

\seealso{
\link{SquIPz} and \link{SIPm}
}
