% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{solve_dAoI}
\alias{solve_dAoI}
\title{Solve the system of differential equations to compute the moments of the AoI over time.}
\usage{
solve_dAoI(h, FoIpar, r = 1/200, tau = 0, Amax = 730, dt = 1, N = 3)
}
\arguments{
\item{h}{the force of infection}

\item{FoIpar}{a FoI trace function}

\item{r}{the clearance rate for a simple infection}

\item{tau}{the cohort birthday}

\item{Amax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}

\item{N}{The total number of moments to compute}
}
\value{
a data.frame with the orbits
}
\description{
Solve the system of differential equations to compute the moments of the AoI over time.
}
