% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{solve_dAoYda}
\alias{solve_dAoYda}
\title{Solve the system of differential equations to compute the approximate moments of the AoY over time.}
\usage{
solve_dAoYda(h, FoIpar, r = 1/200, tau = 0, Amax = 730, dt = 1, n = 8)
}
\arguments{
\item{h}{the force of infection}

\item{FoIpar}{a FoI trace function}

\item{r}{the clearance rate for a simple infection}

\item{tau}{the cohort birthday}

\item{Amax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}

\item{n}{The number of terms to use in \eqn{\phi(r,m)}}
}
\value{
a \link{data.frame} describing the orbits
}
\description{
Solve the system of differential equations to compute the approximate moments of the AoY over time.
}
