% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{solve_dpda}
\alias{solve_dpda}
\title{Solve a system of differential equations to compute the true PR and the MoI}
\usage{
solve_dpda(h, FoIpar, r = 1/200, tau = 0, Amax = 730, dt = 1)
}
\arguments{
\item{h}{a scaling factor on the FoI}

\item{FoIpar}{parameters that define an FoI function}

\item{r}{the clearance rate for a simple infection}

\item{tau}{the cohort birthday}

\item{Amax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}
}
\value{
a \link{data.frame} with the orbits
}
\description{
Solve a system of differential equations to compute the true PR and the MoI
}
