% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPm.R
\name{solve_SIPm}
\alias{solve_SIPm}
\title{Solve the queuing model \eqn{M/M/\infty}}
\usage{
solve_SIPm(
  h,
  FoIpar,
  tau = 0,
  r = 1/200,
  rho = 0.2,
  sigma = 1/365,
  xi = 1/365,
  eta = 1/25,
  mu = 0,
  H = 1000,
  Amax = 730,
  dt = 1
)
}
\arguments{
\item{h}{the force of infection}

\item{FoIpar}{\eqn{h_\tau(a)}, a \link{list} formatted to compute \link{FoI}}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{rho}{the fraction of incident cases that gets treted}

\item{sigma}{treatment rate for infected individuals}

\item{xi}{background drug taking}

\item{eta}{loss of chemoprotection}

\item{mu}{population death rate}

\item{H}{population size}

\item{Amax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}
}
\value{
a \link{list} with the orbits by name
}
\description{
A wrapper to solve the queuing model \eqn{M/M/\infty} (see \link{dMoIda}).

The function automatically sets the maximum MoI to be computed, and it
sets initial conditions. The equations are solved using \link[deSolve:ode]{deSolve::ode} and returned at
regular intervals dt from age 0 up to Amax (in days).
}
\seealso{
\link{dMoIda}
}
