% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquIP.R
\name{solve_SquIP}
\alias{solve_SquIP}
\title{Solve \link{SquIP}}
\usage{
solve_SquIP(
  h,
  FoIpar,
  tau = 0,
  r = 1/200,
  rho = 0.2,
  sigma = 1/365,
  xi = 1/365,
  eta = 1/25,
  mu = 0,
  H = 1000,
  Amax = 730,
  da = 1,
  N = NULL
)
}
\arguments{
\item{h}{the force of infection}

\item{FoIpar}{\eqn{h_\tau(a)}, a \link{list} formatted to compute \link{FoI}}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{rho}{the fraction of incident cases that gets treted}

\item{sigma}{treatment rate for infected individuals}

\item{xi}{background drug taking}

\item{eta}{loss of chemoprotection}

\item{mu}{population death rate}

\item{H}{cohort size}

\item{Amax}{The maximum runtime (in days)}

\item{da}{The output interval (in days)}

\item{N}{truncation parameter (maximum MoI): if \code{NULL}, then set by rule}
}
\value{
a named \link{list} of parsed outputs
}
\description{
This function solves the model \link{SquIP} using \link[deSolve:ode]{deSolve::ode}. It is a wrapper around the derivatives function \link{dSquIP}. It does the following:
\itemize{
\item \code{N} --- set a value for the truncation parameter, if no value was passed
\item \code{y} --- set up the the initial values vector as a named vector
\item \code{times} --- set up a mesh over the independent variable \eqn{a} (age, in days)
\item \code{parms} --- set up a named vector with the parameter values
}

After solving the equations, it parses the outputs, and returns the values of the dependent variables by name. The wrapper
}
\section{Parased Outputs}{


After solving the equations, the solutions are parsed and returned as a named list:
\itemize{
\item \code{age} --- the ages at which the dependent variables were output
\item \code{S} --- Susceptible, a vector
\item \code{P} --- Chemoprotected, a vector
\item \code{Ii} --- Infected \eqn{\times} MoI, a matrix
\item \code{H} --- Cohort population size
\item \code{m_1} --- The hybrid variable \eqn{m_1}, mean MoI
\item \code{m_2} --- The hybrid variable \eqn{m_2}, the second moment of the MoI
\item \code{m1}--- The mean MoI, computed from \code{Ii}
\item \code{m2}--- The second moment of the MoI, computed from \code{Ii}
\item \code{x} --- The prevalence of infection: \eqn{x = (H-P-S)/H}
\item \code{F_r} --- The rate of natural clearance: \eqn{r I_1}
\item \code{N} --- The truncation parameter
\item \code{out} --- The matrix returned by \code{deSolve}
}
}

\seealso{
\link{SquIP}
}
