% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMinfinity.R
\name{solveMMinfty}
\alias{solveMMinfty}
\title{Solve the queuing model \eqn{M/M/\infty}}
\usage{
solveMMinfty(h, FoIpar, r = 1/200, tau = 0, Amax = 730, dt = 1)
}
\arguments{
\item{h}{the force of infection}

\item{FoIpar}{\eqn{h_\tau(a)}, a \link{list} formatted to compute \link{FoI}}

\item{r}{the clearance rate for a simple infection}

\item{tau}{the cohort birthday}

\item{Amax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}
}
\value{
a \link{list} with the orbits by name
}
\description{
A wrapper to solve the queuing model \eqn{M/M/\infty} (see \link{dMoIda}).

The function automatically sets the maximum MoI to be computed, and it
sets initial conditions. The equations are solved using \link[deSolve:ode]{deSolve::ode} and returned at
regular intervals dt from age 0 up to Amax (in days).
}
\seealso{
\link{dMoIda}
}
