% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_dynamics.R
\name{zda}
\alias{zda}
\title{Compute infection density in a cohort of humans, \eqn{z_\tau(\alpha, a |h)}}
\usage{
zda(alpha, a, FoIpar, tau = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{alpha}{the age of an infection, \eqn{\alpha}}

\item{a}{host age}

\item{FoIpar}{a compound \link{list} to compute \eqn{h_\tau(a)}}

\item{tau}{the host cohort's birthday}

\item{hhat}{scaling parameter for \link{FoI}}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
Given a function describing the FoI (\eqn{h_\tau(a)}), and a parameter
describing the clearance rate of infections (\eqn{r}),
the density of parasite clones of age \eqn{\alpha} distributed among a cohort of humans of
age \eqn{a} is \deqn{z_\tau(\alpha, a) = e^{-r \alpha} h_\tau(a-\alpha)}
}
\seealso{
\link{FoI}
}
